trying new idea for multi file explore
